Workaround the icon pack issue on adaptive apps.
On the previous icon pack commit, there was an issue regarding
icon packs and adaptive icons. When an app that got an icon
targeted as SDK +26 (adaptive-ready) and got a custom icon from
an icon pack, the icon would take the new icon as foreground and
it'll take the generic white adaptive background.

Now, with this commit, if the icon packs supports the respective
icon, it'll apply just the icon image instead of doing that.

It's not the best approach, but it's a quick workaround until there's
a better method.

* [StatiXOS]
  - Update for Android 10 by removing code to generate the adaptive
    drawable in BaseIconFactory, going along with the way we do icon
    packs already.

Before: https://i.imgur.com/nlqBVlQ.png?1
After: https://i.imgur.com/HIkcg5D.png?1

Change-Id: Ie639d056413a13897acd2dc1eeb6c5632a0aa645